updated nr task mgr type, and refactor planning stage & flink impl to better divide responsibilities

+++ src/flink/plan/benchmark.rs
@@ -13,7 +13,7 @@ use crate::flink::MetricCatalog;
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct BenchmarkRange {
-    pub nr_task_managers: u8,
+    pub nr_task_managers: u16,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(default)]
@@ -33,7 +33,7 @@ impl BenchmarkRange {
         Self::new(b.nr_task_managers, None, Some(b.records_out_per_sec))
     }
 
-    pub fn new(nr_task_managers: u8, lo_rate: Option<RecordsPerSecond>, hi_rate: Option<RecordsPerSecond>) -> Self {
+    pub fn new(nr_task_managers: u16, lo_rate: Option<RecordsPerSecond>, hi_rate: Option<RecordsPerSecond>) -> Self {
         Self { nr_task_managers, lo_rate, hi_rate }
     }
 }
@@ -165,7 +165,7 @@ impl RelativeEq for BenchmarkRange {
 #[serde_as]
 #[derive(Debug, Copy, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Benchmark {
-    pub nr_task_managers: u8,
+    pub nr_task_managers: u16,
     pub records_out_per_sec: RecordsPerSecond,
 }
 
@@ -176,7 +176,7 @@ impl fmt::Display for Benchmark {
 }
 
 impl Benchmark {
-    pub fn new(nr_task_managers: u8, records_out_per_sec: RecordsPerSecond) -> Self {
+    pub fn new(nr_task_managers: u16, records_out_per_sec: RecordsPerSecond) -> Self {
         Self { nr_task_managers, records_out_per_sec }
     }
 }
@@ -250,7 +250,7 @@ impl TryFrom<TelemetryValue> for Benchmark {
         if let TelemetryValue::Table(rep) = telemetry {
             let nr_task_managers = rep
                 .get(T_NR_TASK_MANAGERS)
-                .map(|v| u8::try_from(v.clone()))
+                .map(|v| u16::try_from(v.clone()))
                 .ok_or(PlanError::DataNotFound(T_NR_TASK_MANAGERS.to_string()))??;
 
             let records_out_per_sec = rep
